Moving functions around for circular dep
Preventing circular dependencies across files.
